Move ledgers sync-schema to separate file
